Split assets functions for readability
This allows developers working to more easily correlate handlers when debugging locally. Production servers should route the paths via whatever web server runs in front and _not_ run in testing mode.
